refactor: split lib/api.ts (1364 lines) into focused modules
- lib/api/client.ts    — shared fetch, auth headers, config (89 lines)
- lib/api/portal.ts    — calendar, announcements, rules, newsletters (149 lines)
- lib/api/members.ts   — members + activities with mock fallback (198 lines)
- lib/api/resources.ts — portal resources (42 lines)
- lib/api/public-content.ts — news, training, resources, pages, officials, exec team (116 lines)
- lib/api/auth.ts      — Supabase Auth admin API + types (119 lines)
- lib/api/evaluations.ts — evaluations API + types (97 lines)
- lib/api/index.ts     — barrel re-export (all @/lib/api imports unchanged)
